@@ -4,7 +4,7 @@
 
 	<groupId>fi.otavanopisto.muikku</groupId>
 	<artifactId>parent</artifactId>
-	<version>1.1.41-SNAPSHOT</version>
+	<version>1.1.42-SNAPSHOT</version>
 	<packaging>pom</packaging>
 	<name>Muikku Parent</name>
 
@@ -77,7 +77,7 @@
     <module>muikku-timed-notifications-plugin-persistence</module>
 		<module>muikku</module>
     <module>muikku-developer-tools</module>
-		<module>muikku-atests</module>
+		<!-- <module>muikku-atests</module> -->
     <module>muikku-atests-plugin</module>
     <module>muikku-atests-plugin-rest-model</module>
 	</modules>
@@ -85,8 +85,8 @@
 	<properties>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 		<maven.build.timestamp.format>yyyyMMdd'T'HHmmss</maven.build.timestamp.format>
-		<maven.compiler.source>1.7</maven.compiler.source>
-		<maven.compiler.target>1.7</maven.compiler.target>
+		<maven.compiler.source>1.8</maven.compiler.source>
+		<maven.compiler.target>1.8</maven.compiler.target>
 		<hibernate-jpamodelgen.version>4.3.6.Final</hibernate-jpamodelgen.version>
     <findbugs.skip>true</findbugs.skip>
 	</properties>
@@ -96,7 +96,7 @@
 			<dependency>
 				<groupId>fi.otavanopisto.muikku</groupId>
 				<artifactId>bom</artifactId>
-				<version>1.1.41-SNAPSHOT</version>
+				<version>1.1.42-SNAPSHOT</version>
 				<type>pom</type>
 				<scope>import</scope>
 			</dependency>
