@@ -16,7 +16,7 @@ import org.joda.time.DateTime;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.SerializationFeature;
-import com.fasterxml.jackson.datatype.joda.JodaModule;
+import com.fasterxml.jackson.datatype.jsr310.JSR310Module;
 
 import fi.otavanopisto.muikku.AbstractPyramusMocks;
 import fi.otavanopisto.muikku.TestUtilities;
@@ -70,7 +70,7 @@ public class PyramusMocks extends AbstractPyramusMocks {
   }
   
   private static void loginMock(Long userId, String email, String firstName, String lastName) throws JsonProcessingException {
-    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JodaModule()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
+    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JSR310Module()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
     
     stubFor(get(urlEqualTo("/dnm")).willReturn(aResponse().withHeader("Content-Type", "application/json").withBody("").withStatus(204)));
 
@@ -106,7 +106,7 @@ public class PyramusMocks extends AbstractPyramusMocks {
   }
   
   public static void personsPyramusMocks() throws Exception {
-    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JodaModule()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
+    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JSR310Module()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
     
     Map<String, String> variables = null;
     List<String> tags = null;
@@ -496,7 +496,7 @@ public class PyramusMocks extends AbstractPyramusMocks {
   
   @SuppressWarnings("unused")
   private static StaffMember mockStaffMember(Long personId, Long staffMemberId, String firstName, String lastName, String email, UserRole role, List<String> tags, Map<String, String> variables, List<String> payloads) throws JsonProcessingException {
-    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JodaModule()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
+    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JSR310Module()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
 
     StaffMember staffMember = new StaffMember(staffMemberId, personId, null, firstName, lastName, null, 
         role, tags, variables);
@@ -534,7 +534,7 @@ public class PyramusMocks extends AbstractPyramusMocks {
   }
   
   private static void mockStudyProgrammes() throws JsonProcessingException {
-    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JodaModule()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
+    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JSR310Module()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
     
     StudyProgrammeCategory spc = new StudyProgrammeCategory(1l, "All Study Programmes", 1l, false);
     StudyProgramme sp = new StudyProgramme(1l, "test", "Test Study Programme", 1l, false);
@@ -572,7 +572,7 @@ public class PyramusMocks extends AbstractPyramusMocks {
   }
   
   public static void workspacePyramusMock(Long id, String name, String description) throws Exception {
-    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JodaModule()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
+    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JSR310Module()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
     DateTime created = new DateTime(1990, 2, 2, 0, 0, 0, 0);
     DateTime begin = new DateTime(2000, 1, 1, 0, 0, 0, 0);
     DateTime end = new DateTime(2050, 1, 1, 0, 0, 0, 0);
@@ -620,7 +620,7 @@ public class PyramusMocks extends AbstractPyramusMocks {
   }
 
   public static void studentGroupsMocks() throws JsonProcessingException {
-    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JodaModule()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
+    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JSR310Module()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
     DateTime created = new DateTime(1990, 2, 2, 0, 0, 0, 0);
     DateTime begin = new DateTime(2000, 1, 1, 0, 0, 0, 0);
     DateTime lastmodified = new DateTime(2000, 1, 1, 0, 0, 0, 0);
@@ -636,7 +636,7 @@ public class PyramusMocks extends AbstractPyramusMocks {
   }
   
   public static void adminMock() throws Exception {
-    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JodaModule()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);    
+    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JSR310Module()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);    
     
     Map<String, String> variables = null;
     List<String> tags = null;
@@ -740,7 +740,7 @@ public class PyramusMocks extends AbstractPyramusMocks {
   }
   
   private static Person mockPerson(Long personId, DateTime birthday, String socialSecurityNumber, Sex sex, Long defaultUserId) throws JsonProcessingException {
-    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JodaModule()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
+    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JSR310Module()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
     
     Person person = new Person(personId, birthday, socialSecurityNumber, sex, false, "empty", defaultUserId);
     String personJson = objectMapper.writeValueAsString(person);
@@ -755,7 +755,7 @@ public class PyramusMocks extends AbstractPyramusMocks {
   }
   
   public static void mockCommons() throws JsonProcessingException {
-    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JodaModule()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
+    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JSR310Module()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
 
     Subject subject = new Subject((long) 1, "tc_11", "Test course", (long) 1, false);
     String subjectJson = objectMapper.writeValueAsString(subject);
@@ -856,7 +856,7 @@ public class PyramusMocks extends AbstractPyramusMocks {
   }
 
   public static void mockCourseTypes() throws JsonProcessingException {
-    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JodaModule()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
+    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JSR310Module()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
     
     fi.otavanopisto.pyramus.rest.model.CourseType courseType = new fi.otavanopisto.pyramus.rest.model.CourseType((long) 1, "Nonstop", false);
     CourseType[] courseTypeArray = { courseType };
@@ -879,7 +879,7 @@ public class PyramusMocks extends AbstractPyramusMocks {
   }
   
   public static void mockContactTypes() throws JsonProcessingException {
-    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JodaModule()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
+    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JSR310Module()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
     
     ContactType contactType = new ContactType((long)1, "Koti", false, false);
     ContactType[] contactTypes = { contactType };
@@ -901,7 +901,7 @@ public class PyramusMocks extends AbstractPyramusMocks {
   public static void mockAssessedStudent1Workspace1(MockCourseStudent courseStudent, Long assessorId) throws JsonProcessingException {
     // TODO: Move to new mocker.
     
-    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JodaModule()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
+    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JSR310Module()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
 //    CourseStudent courseStudent = new CourseStudent(3l, 1l, 1l, new DateTime(2010, 2, 2, 0, 0, 0, 0), false, null, null, null, null, null);
     DateTime assessmentCreated = new DateTime(2015, 2, 2, 0, 0, 0, 0);
     CourseAssessment courseAssessment = new CourseAssessment(1l, courseStudent.getId(), 1l, 1l, assessorId, assessmentCreated, "Test evaluation.");
