@@ -14,7 +14,7 @@ import javax.inject.Inject;
 
 import org.apache.commons.collections.CollectionUtils;
 import org.apache.commons.lang3.StringUtils;
-import org.threeten.bp.ZonedDateTime;
+import java.time.OffsetDateTime;
 
 import fi.otavanopisto.muikku.calendar.Calendar;
 import fi.otavanopisto.muikku.calendar.CalendarEvent;
@@ -117,7 +117,7 @@ public class CalendarController {
     return provider.findEvent(calendar, eventId);
   }
 
-  public List<fi.otavanopisto.muikku.calendar.CalendarEvent> listCalendarEvents(UserCalendar userCalendar, ZonedDateTime timeMin, ZonedDateTime timeMax) throws CalendarServiceException {
+  public List<fi.otavanopisto.muikku.calendar.CalendarEvent> listCalendarEvents(UserCalendar userCalendar, OffsetDateTime timeMin, OffsetDateTime timeMax) throws CalendarServiceException {
     CalendarServiceProvider provider = getCalendarServiceProvider(userCalendar.getCalendarProvider());
 
     Calendar calendar = provider.findCalendar(userCalendar.getCalendarId());
