@@ -578,7 +578,8 @@ public class CommunicatorRESTService extends PluginRESTService {
           user.getLastName(), 
           user.getStudyProgrammeName(),
           hasPicture,
-          user.hasEvaluationFees());
+          user.hasEvaluationFees(),
+          user.getCurriculumIdentifier());
       
       return Response.ok(
         result
@@ -610,7 +611,8 @@ public class CommunicatorRESTService extends PluginRESTService {
           user.getLastName(), 
           user.getStudyProgrammeName(),
           hasPicture,
-          user.hasEvaluationFees());
+          user.hasEvaluationFees(),
+          user.getCurriculumIdentifier());
     
       return Response.ok(
         result
