@@ -1,6 +1,6 @@
 package fi.otavanopisto.muikku.plugins.schooldatalocal.entities;
 
-import org.threeten.bp.ZonedDateTime;
+import java.time.OffsetDateTime;
 
 import fi.otavanopisto.muikku.plugins.schooldatalocal.LocalUserSchoolDataController;
 import fi.otavanopisto.muikku.schooldata.entity.AbstractUser;
@@ -10,9 +10,9 @@ public class LocalUserImpl  extends AbstractUser implements User {
 
   public LocalUserImpl(String identifier, String firstName, String lastName, String displayName,
       String studyProgrammeName, String nationality, String language, String municipality, String school,
-      ZonedDateTime studyStartDate, ZonedDateTime studyEndDate, ZonedDateTime studyTimeEnd) {
+      OffsetDateTime studyStartDate, OffsetDateTime studyEndDate, OffsetDateTime studyTimeEnd, String curriculumIdentifier) {
     super(identifier, firstName, lastName, displayName, studyProgrammeName, nationality, language, municipality, school,
-        studyStartDate, studyEndDate, studyTimeEnd, false, false, false, false, false);
+        studyStartDate, studyEndDate, studyTimeEnd, false, false, false, false, false, curriculumIdentifier);
   }
 
   @Override
