@@ -16,12 +16,33 @@
         if (err) {
           $('.notification-queue').notificationQueue('notification', 'error', err);
         } else {
-          this._hasEvaluationFees = results[0];
-          var educationTypes = results[1];
+          this._hasEvaluationFees = results.evaluationFees;
+          var educationTypes = results.educationTypes;
+          var curriculums = results.curriculums;
+          var defaultCurriculum = results.userCurriculum;
+
+          if (defaultCurriculum)
+            this._curriculums = [ defaultCurriculum ];
           
-          renderDustTemplate('coursepicker/coursepicker_educationtype_filters.dust', { educationTypes: educationTypes }, $.proxy(function (text) {
-            this.element.find('.cp-filters').html(text);
-          }, this));
+          async.series([
+              $.proxy(function (callback) {
+                renderDustTemplate('coursepicker/coursepicker_educationtype_filters.dust', { educationTypes: educationTypes }, $.proxy(function (text) {
+                  callback(null, $(text));
+                }, this))
+              }, this),
+    
+              $.proxy(function (callback) {
+                renderDustTemplate('coursepicker/coursepicker_curriculum_filters.dust', { curriculums: curriculums, defaultCurriculum: defaultCurriculum }, $.proxy(function (text) {
+                  callback(null, $(text));
+                }, this))
+                callback();
+              }, this)
+            ],
+            $.proxy(function (err, results) {
+              this.element.find('.cp-filters').append(results[0]);
+              this.element.find('.cp-filters').append(results[1]);
+            }, this)
+          );
           
           this.element.on('change', ".cp-category-dropdown", $.proxy(this._onCategoryChange, this));
           this.element.on('keyup', "input[name='coursePickerSearch']", $.proxy(this._onSearchKeyUp, this));
@@ -32,6 +53,7 @@
           this.element.on("click", ".cp-course-tour-button", $.proxy(this._onTourButtonClick, this));
           this.element.on("click", ".cp-course-attend-button", $.proxy(this._onAttendButtonClick, this));
           this.element.on("click", ".cp-education-type-filter", $.proxy(this._onEducationTypeFilterClick, this));
+          this.element.on("click", ".cp-curriculum-filter", $.proxy(this._onCurriculumFilterClick, this));
           
           this._reloadWorkspaces();
         }
@@ -63,10 +85,23 @@
       this._reloadWorkspaces();
     },
     
+    curriculums: function (curriculums) {
+      this._curriculums = curriculums||[];
+      
+      this._reloadWorkspaces();
+    },
+    
     _load: function (callback) {
-      async.parallel([this._createResolveEvaluationFees(), this._createEducationTypesLoad()], function (err, results) {
-        callback(err, results);
-      }); 
+      async.parallel({
+          evaluationFees: this._createResolveEvaluationFees(), 
+          educationTypes: this._createEducationTypesLoad(),
+          curriculums: this._createCurriculumsLoad(),
+          userCurriculum: this._createFindUserCurriculum()
+        }, 
+        function (err, results) {
+          callback(err, results);
+        }
+      ); 
     },
     
     _createResolveEvaluationFees: function () {
@@ -91,6 +126,28 @@
       }, this);
     },
     
+    _createCurriculumsLoad: function () {
+      return $.proxy(function (callback) {
+        mApi().coursepicker.curriculums
+          .read()
+          .callback(callback);
+      }, this);
+    },
+    
+    _createFindUserCurriculum: function () {
+      return $.proxy(function (callback) {
+        if (!MUIKKU_LOGGED_USER) {
+          callback(null, false);
+        } else {
+          mApi().user.users.basicinfo
+            .read(MUIKKU_LOGGED_USER)
+            .callback($.proxy(function (err, basicInfo) {
+              callback(err, basicInfo && basicInfo.curriculumIdentifier);
+            }, this));
+        }
+      }, this);
+    },
+    
     _reloadWorkspaces: function () {
       this._firstResult = 0;
       this.element.find('#coursesList').empty();
@@ -109,7 +166,8 @@
         search: this._search,
         myWorkspaces: this._categoryId == 'te' ||  this._categoryId == 'my',
         includeUnpublished: this._categoryId == 'te',
-        educationTypes: this._educationTypes
+        educationTypes: this._educationTypes,
+        curriculums: this._curriculums
       };
       
       mApi().coursepicker.workspaces
@@ -252,6 +310,21 @@
         var typeId = filterElement.attr('data-id');
         this.educationTypes([typeId]);
       }
+    },
+    _onCurriculumFilterClick: function (event) {
+      $(event.target).closest("li").toggleClass("selected");
+      var curriculumOptions = $(event.target).closest('.cp-filters-ul').find("li");
+      var selectedCurriculums = [];
+      
+      curriculumOptions.each(function () {
+        var curriculumLi = $(this);
+        if (curriculumLi.hasClass('selected')) {
+          var curriculumId = curriculumLi.find(".cp-curriculum-filter").attr("data-id");
+          selectedCurriculums.push(curriculumId);
+        }
+      });
+
+      this.curriculums(selectedCurriculums);
     }
   });
 
