@@ -47,7 +47,8 @@
     }, this));
   }
   
-  $(document).on('click', '.bt-forgotpassword', function (event, data) {
+  $(document).on('click', '.forgotpassword-link', function (event, data) {
+	event.preventDefault();
     confirmForgotpasswordRequest();
   });
 
