@@ -1,13 +1,13 @@
 package fi.otavanopisto.muikku.schooldata.entity;
 
-import org.threeten.bp.ZonedDateTime;
+import java.time.OffsetDateTime;
 
 public abstract class AbstractUser implements User {
   
   public AbstractUser(String identifier, String firstName, String lastName, String displayName,
       String studyProgrammeName, String nationality, String language, String municipality, String school,
-      ZonedDateTime studyStartDate, ZonedDateTime studyEndDate, ZonedDateTime studyTimeEnd, boolean hidden, 
-      boolean startedStudies, boolean finishedStudies, boolean active, boolean evaluationFees) {
+      OffsetDateTime studyStartDate, OffsetDateTime studyEndDate, OffsetDateTime studyTimeEnd, boolean hidden, 
+      boolean startedStudies, boolean finishedStudies, boolean active, boolean evaluationFees, String curriculumIdentifier) {
     super();
     this.identifier = identifier;
     this.firstName = firstName;
@@ -26,6 +26,7 @@ public abstract class AbstractUser implements User {
     this.startedStudies = startedStudies;
     this.finishedStudies = finishedStudies;
     this.evaluationFees = evaluationFees;
+    this.curriculumIdentifier = curriculumIdentifier;
   }
 
   @Override
@@ -105,29 +106,29 @@ public abstract class AbstractUser implements User {
   }
 
   @Override
-  public ZonedDateTime getStudyStartDate() {
+  public OffsetDateTime getStudyStartDate() {
     return studyStartDate;
   }
 
-  public void setStudyStartDate(ZonedDateTime studyStartDate) {
+  public void setStudyStartDate(OffsetDateTime studyStartDate) {
     this.studyStartDate = studyStartDate;
   }
   
   @Override
-  public ZonedDateTime getStudyEndDate() {
+  public OffsetDateTime getStudyEndDate() {
     return this.studyEndDate;
   }
   
-  public void setStudyEndDate(ZonedDateTime studyEndDate) {
+  public void setStudyEndDate(OffsetDateTime studyEndDate) {
     this.studyEndDate = studyEndDate;
   }
 
   @Override
-  public ZonedDateTime getStudyTimeEnd() {
+  public OffsetDateTime getStudyTimeEnd() {
     return studyTimeEnd;
   }
 
-  public void setStudyTimeEnd(ZonedDateTime studyTimeEnd) {
+  public void setStudyTimeEnd(OffsetDateTime studyTimeEnd) {
     this.studyTimeEnd = studyTimeEnd;
   }
 
@@ -172,6 +173,15 @@ public abstract class AbstractUser implements User {
     return evaluationFees;
   }
 
+  @Override
+  public String getCurriculumIdentifier() {
+    return curriculumIdentifier;
+  }
+
+  public void setCurriculumIdentifier(String curriculumIdentifier) {
+    this.curriculumIdentifier = curriculumIdentifier;
+  }
+
   private String identifier;
   private String firstName;
   private String lastName;
@@ -181,12 +191,13 @@ public abstract class AbstractUser implements User {
   private String language;
   private String municipality;
   private String school;
-  private ZonedDateTime studyStartDate;
-  private ZonedDateTime studyEndDate;
-  private ZonedDateTime studyTimeEnd;
+  private OffsetDateTime studyStartDate;
+  private OffsetDateTime studyEndDate;
+  private OffsetDateTime studyTimeEnd;
   private boolean hidden;
   private boolean startedStudies;
   private boolean finishedStudies;
   private boolean active;
   private boolean evaluationFees;
+  private String curriculumIdentifier;
 }
