@@ -2,7 +2,7 @@ package fi.otavanopisto.muikku.schooldata.entity;
 
 import java.util.Date;
 
-import org.threeten.bp.ZonedDateTime;
+import java.time.OffsetDateTime;
 
 import fi.otavanopisto.muikku.schooldata.SchoolDataIdentifier;
 
@@ -14,7 +14,7 @@ public abstract class AbstractWorkspace implements Workspace {
   public AbstractWorkspace(String identifier, String name, String nameExtension, String viewLink,
       SchoolDataIdentifier workspaceTypeId, String courseIdentifierIdentifier, String description,
       String subjectIdentifier, SchoolDataIdentifier educationTypeIdentifier, Date modified, Double length, String lengthUnitIdentifier,
-      ZonedDateTime beginDate, ZonedDateTime endDate, boolean archived, boolean evaluationFeeApplicable) {
+      OffsetDateTime beginDate, OffsetDateTime endDate, boolean archived, boolean evaluationFeeApplicable, SchoolDataIdentifier curriculumIdentifier) {
     super();
     this.identifier = identifier;
     this.name = name;
@@ -32,6 +32,7 @@ public abstract class AbstractWorkspace implements Workspace {
     this.endDate = endDate;
     this.archived = archived;
     this.evaluationFeeApplicable = evaluationFeeApplicable;
+    this.curriculumIdentifier = curriculumIdentifier;
   }
   
   @Override
@@ -113,22 +114,22 @@ public abstract class AbstractWorkspace implements Workspace {
   }
   
   @Override
-  public ZonedDateTime getBeginDate() {
+  public OffsetDateTime getBeginDate() {
     return beginDate;
   }
   
   @Override
-  public void setBeginDate(ZonedDateTime beginDate) {
+  public void setBeginDate(OffsetDateTime beginDate) {
     this.beginDate = beginDate; 
   }
   
   @Override
-  public ZonedDateTime getEndDate() {
+  public OffsetDateTime getEndDate() {
     return endDate;
   }
   
   @Override
-  public void setEndDate(ZonedDateTime endDate) {
+  public void setEndDate(OffsetDateTime endDate) {
     this.endDate = endDate;
   }
   
@@ -147,6 +148,11 @@ public abstract class AbstractWorkspace implements Workspace {
     return viewLink;
   }
 
+  @Override
+  public SchoolDataIdentifier getCurriculumIdentifier() {
+    return curriculumIdentifier;
+  }
+  
   private String identifier;
   private String name;
   private String nameExtension;
@@ -159,8 +165,9 @@ public abstract class AbstractWorkspace implements Workspace {
   private Date modified;
   private Double length;
   private String lengthUnitIdentifier;
-  private ZonedDateTime beginDate;
-  private ZonedDateTime endDate;
+  private OffsetDateTime beginDate;
+  private OffsetDateTime endDate;
   private boolean archived;
   private boolean evaluationFeeApplicable;
+  private SchoolDataIdentifier curriculumIdentifier;
 }
