@@ -1,6 +1,6 @@
 package fi.otavanopisto.muikku.schooldata.entity;
 
-import org.threeten.bp.ZonedDateTime;
+import java.time.OffsetDateTime;
 
 import fi.otavanopisto.muikku.schooldata.SchoolDataIdentifier;
 
@@ -10,6 +10,6 @@ public interface WorkspaceUser extends SchoolDataEntity {
 	public SchoolDataIdentifier getUserIdentifier();
 	public SchoolDataIdentifier getWorkspaceIdentifier();
 	public SchoolDataIdentifier getRoleIdentifier();
-	public ZonedDateTime getEnrolmentTime();
+	public OffsetDateTime getEnrolmentTime();
 	
 }
