@@ -39,7 +39,6 @@ import org.apache.commons.codec.digest.DigestUtils;
 import org.apache.commons.collections.CollectionUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.lang3.math.NumberUtils;
-import org.threeten.bp.DateTimeUtils;
 
 import fi.otavanopisto.muikku.model.users.EnvironmentRoleArchetype;
 import fi.otavanopisto.muikku.model.users.Flag;
@@ -351,9 +350,9 @@ public class UserRESTService extends AbstractRESTService {
     }
     
     String emailAddress = userEmailEntityController.getUserDefaultEmailAddress(userEntity, true); 
-    Date studyStartDate = user.getStudyStartDate() != null ? DateTimeUtils.toDate(user.getStudyStartDate().toInstant()) : null;
-    Date studyEndDate = user.getStudyEndDate() != null ? DateTimeUtils.toDate(user.getStudyEndDate().toInstant()) : null;
-    Date studyTimeEnd = user.getStudyTimeEnd() != null ? DateTimeUtils.toDate(user.getStudyTimeEnd().toInstant()) : null;
+    Date studyStartDate = user.getStudyStartDate() != null ? Date.from(user.getStudyStartDate().toInstant()) : null;
+    Date studyEndDate = user.getStudyEndDate() != null ? Date.from(user.getStudyEndDate().toInstant()) : null;
+    Date studyTimeEnd = user.getStudyTimeEnd() != null ? Date.from(user.getStudyTimeEnd().toInstant()) : null;
     
     Student student = new Student(
         studentIdentifier.toId(), 
@@ -983,7 +982,7 @@ public class UserRESTService extends AbstractRESTService {
       // TODO: User image
       boolean hasImage = false;
       return Response
-          .ok(new UserBasicInfo(userEntity.getId(), user.getFirstName(), user.getLastName(), user.getStudyProgrammeName(), hasImage, user.hasEvaluationFees()))
+          .ok(new UserBasicInfo(userEntity.getId(), user.getFirstName(), user.getLastName(), user.getStudyProgrammeName(), hasImage, user.hasEvaluationFees(), user.getCurriculumIdentifier()))
           .cacheControl(cacheControl)
           .tag(tag)
           .build();
@@ -1064,8 +1063,8 @@ public class UserRESTService extends AbstractRESTService {
 		
 		String emailAddress = userEmailEntityController.getUserDefaultEmailAddress(userEntity, true); 
 		
-		Date startDate = user.getStudyStartDate() != null ? DateTimeUtils.toDate(user.getStudyStartDate().toInstant()) : null;
-		Date endDate = user.getStudyTimeEnd() != null ? DateTimeUtils.toDate(user.getStudyTimeEnd().toInstant()) : null;
+		Date startDate = user.getStudyStartDate() != null ? Date.from(user.getStudyStartDate().toInstant()) : null;
+		Date endDate = user.getStudyTimeEnd() != null ? Date.from(user.getStudyTimeEnd().toInstant()) : null;
 		
 		return new fi.otavanopisto.muikku.rest.model.User(userEntity.getId(),
 				user.getFirstName(), user.getLastName(), hasImage,
