@@ -1,6 +1,6 @@
 package fi.otavanopisto.muikku.plugins.schooldatapyramus.entities;
 
-import org.threeten.bp.ZonedDateTime;
+import java.time.OffsetDateTime;
 
 import fi.otavanopisto.muikku.plugins.schooldatapyramus.SchoolDataPyramusPluginDescriptor;
 import fi.otavanopisto.muikku.schooldata.SchoolDataIdentifier;
@@ -10,7 +10,7 @@ import fi.otavanopisto.muikku.schooldata.entity.WorkspaceUser;
 public class PyramusWorkspaceUser extends AbstractWorkspaceUser implements WorkspaceUser {
 
 	public PyramusWorkspaceUser(SchoolDataIdentifier identifier, SchoolDataIdentifier userIdentifier, SchoolDataIdentifier workspaceIdentifier,
-      SchoolDataIdentifier roleIdentifier, ZonedDateTime enrolmentTime) {
+      SchoolDataIdentifier roleIdentifier, OffsetDateTime enrolmentTime) {
     super(identifier, userIdentifier, workspaceIdentifier, roleIdentifier, enrolmentTime);
   }
 
