@@ -10,7 +10,7 @@ import javax.inject.Inject;
 import javax.ws.rs.client.Client;
 
 import org.apache.commons.lang3.StringUtils;
-import org.threeten.bp.ZonedDateTime;
+import java.time.OffsetDateTime;
 
 import fi.otavanopisto.muikku.controller.PluginSettingsController;
 import fi.otavanopisto.muikku.plugins.schooldatapyramus.SchoolDataPyramusPluginDescriptor;
@@ -43,12 +43,12 @@ public class SystemAccessTokenProvider {
       AccessToken createdAccessToken = restClient.createAccessToken(client, authCode);
       accessToken = createdAccessToken.getAccessToken();
       refreshToken = createdAccessToken.getRefreshToken();
-      accessTokenExpires = ZonedDateTime.now().plusSeconds(createdAccessToken.getExpiresIn());
+      accessTokenExpires = OffsetDateTime.now().plusSeconds(createdAccessToken.getExpiresIn());
     } else {
       if ((accessTokenExpires == null) || (System.currentTimeMillis() > accessTokenExpires.toInstant().toEpochMilli())) {
         AccessToken refreshedAccessToken = restClient.refreshAccessToken(client, refreshToken);
         accessToken = refreshedAccessToken.getAccessToken();
-        accessTokenExpires = ZonedDateTime.now().plusSeconds(refreshedAccessToken.getExpiresIn() - EXPIRE_SLACK);
+        accessTokenExpires = OffsetDateTime.now().plusSeconds(refreshedAccessToken.getExpiresIn() - EXPIRE_SLACK);
       }
     }
 
@@ -57,6 +57,6 @@ public class SystemAccessTokenProvider {
   
   private String accessToken;
   private String refreshToken;
-  private ZonedDateTime accessTokenExpires;
+  private OffsetDateTime accessTokenExpires;
   private String authCode;
 }
